Refactor: Remove generic type `VM: VMBinding` from scheduler types
