+++ src/lib.rs
@@ -10,6 +10,7 @@
 #![feature(arbitrary_self_types)]
 #![feature(associated_type_defaults)]
 #![feature(specialization)]
+#![feature(trait_alias)]
 
 //! Memory Management ToolKit (MMTk) is a portable and high performance memory manager
 //! that includes various garbage collection algorithms and provides clean and efficient
