+++ src/mmtk.rs
@@ -39,7 +39,7 @@ pub struct MMTK<VM: VMBinding> {
     pub sftmap: &'static SFTMap,
     pub reference_processors: ReferenceProcessors,
     pub options: Arc<UnsafeOptionsWrapper>,
-    pub scheduler: Arc<Scheduler<VM>>,
+    pub scheduler: Arc<Scheduler<Self>>,
 
     inside_harness: AtomicBool,
 }
