+++ src/plan/controller_collector_context.rs
@@ -26,7 +26,7 @@ pub struct ControllerCollectorContext<VM: VMBinding> {
     request_condvar: Condvar,
 
     mmtk: &'static MMTK<VM>,
-    pub scheduler: Arc<Scheduler<VM>>,
+    pub scheduler: Arc<MMTkScheduler<VM>>,
     request_flag: AtomicBool,
     phantom: PhantomData<VM>,
 }
