+++ src/plan/semispace/mod.rs
@@ -15,6 +15,7 @@ pub use self::tracelocal::SSTraceLocal as SelectedTraceLocal;
 
 
 use crate::scheduler::*;
+use crate::scheduler::gc_works::*;
 use crate::util::{Address, ObjectReference, OpaquePointer};
 use crate::vm::VMBinding;
 use std::marker::PhantomData;
@@ -68,10 +69,10 @@ impl <VM: VMBinding> ProcessEdgesWork for SSProcessEdges<VM> {
             return object;
         }
         if self.plan().tospace().in_space(object) {
-            return self.plan().tospace().trace_object(self, object, global::ALLOC_SS, self.worker().context());
+            return self.plan().tospace().trace_object(self, object, global::ALLOC_SS, self.worker().local());
         }
         if self.plan().fromspace().in_space(object) {
-            return self.plan().fromspace().trace_object(self, object, global::ALLOC_SS, self.worker().context());
+            return self.plan().fromspace().trace_object(self, object, global::ALLOC_SS, self.worker().local());
         }
         self.plan().common.trace_object(self, object)
     }
